Problem definition and scope
The rise in property prices in Ireland is directly affecting the ability of citizens to buy a home and affecting the standard of living of citizens. This study aims to identify the factors affecting housing prices based on real data and to make decisions such as buying or selling, and to analyze future price trends. The scope of the project is to study property price data in Ireland between 2015 and 2025.